requirements.txt
requests==2.21.0
#Flask==1.0.2 #Flask is already installed on the image
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
flask-sqlalchemy
